+++ pyxel/models/charge_generation/__init__.py
@@ -22,3 +22,4 @@ from .charge_deposition import charge_deposition
 from .charge_deposition import charge_deposition_in_mct
 from .wavelength_qe import apply_qe_curve, conversion_with_3d_qe_map
 from .exponential_qe import exponential_qe
+from .avalanche import avalanche_gain